+++ main.js
@@ -214,11 +214,14 @@ export function packArrayTo(values, theType, buffer, index=0) {
  * @param {!Uint8Array|!Array<!number>} buffer The byte buffer.
  * @param {!Object} theType The type definition.
  * @param {number=} index The buffer index to read.
- * @return {number}
+ * @return {number|undefined}
  * @throws {Error} If the type definition is not valid
  */
 export function unpack(buffer, theType, index=0) {
   setUp_(theType);
+  if ((theType.offset + index) > buffer.length) {
+    throw Error('Bad buffer length.');
+  }
   if (theType.be) {
     endianness(buffer, theType.offset, index, index + theType.offset);
   }
@@ -256,17 +259,12 @@ export function unpackArray(buffer, theType, index=0, end=buffer.length) {
  * @throws {Error} If the type definition is not valid
  */
 export function unpackArrayTo(buffer, theType, output, index=0, end=buffer.length) {
-  setUp_(theType);
   while ((end - index) % theType.offset) {
-    end--;
-  }
-  if (theType.be) {
-    endianness(buffer, theType.offset);
+      end--;
   }
   for (let i = 0; index < end; index += theType.offset, i++) {
-    output[i] = reader_(buffer, index);
-  }
-  if (theType.be) {
-    endianness(buffer, theType.offset);
+    //if ((theType.offset + index - 1) < buffer.length) {
+    output[i] = unpack(buffer, theType, index);
+    //}
   }
 }